% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\docType{methods}
\name{addIdentInfo}
\alias{addIdentInfo}
\title{Add identification result into metaXpara object}
\usage{
addIdentInfo(para, file, ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{file}{The file name which contains the identification result}

\item{...}{Other argument}
}
\value{
A metaXpara object.
}
\description{
Add identification result into metaXpara object
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
