% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{addValueNorm<-}
\alias{addValueNorm<-}
\title{addValueNorm}
\usage{
addValueNorm(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{An object of metaXpara}
}
\value{
An object of metaXpara
}
\description{
addValueNorm
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
addValueNorm(para) <- para
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
