% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{autoRemoveOutlier}
\alias{autoRemoveOutlier}
\title{Automatically detect outlier samples}
\usage{
autoRemoveOutlier(para, outTol = 1.2, pcaMethod = "svdImpute",
  valueID = "valueNorm", scale = "none", center = FALSE, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{outTol}{A factor to define the outlier tolerance, default is 1.2}

\item{pcaMethod}{See \code{\link{pca}} in \pkg{pcaMethods}}

\item{valueID}{The name of the column which will be used}

\item{scale}{Scaling, see \code{\link{pca}} in \pkg{pcaMethods}}

\item{center}{Centering, see \code{\link{pca}} in \pkg{pcaMethods}}

\item{...}{Additional parameter}
}
\value{
The name of outlier samples
}
\description{
Automatically detect outlier samples
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
rs <- autoRemoveOutlier(para,valueID="value")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

