% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPLSDA.R
\name{bootPLSDA}
\alias{bootPLSDA}
\title{Fit predictive models for PLS-DA}
\usage{
bootPLSDA(x, y, ncomp = 2, sample = NULL, test = NULL, split = 0,
  method = "repeatedcv", repeats = 250, number = 7, ...)
}
\arguments{
\item{x}{An object where samples are in rows and features are in columns. 
This could be a simple matrix, data frame.}

\item{y}{A numeric or factor vector containing the outcome for each sample.}

\item{ncomp}{The maximal number of component for PLS-DA}

\item{sample}{A vector contains the sample used for the model}

\item{test}{The data set (data.frame) for testing. If the data contains a 
column with the name "class", this column is the sample class.}

\item{split}{Whether split the data as train and test set. Default is 0 
which indicates not split the data.}

\item{method}{The resampling method: boot, boot632, cv, repeatedcv, LOOCV, 
LGOCV (for repeated training/test splits), none (only fits one model to the 
entire training set), oob (only for random forest, bagged trees, bagged 
earth, bagged flexible discriminant analysis, or conditional tree forest 
models), "adaptive_cv", "adaptive_boot" or "adaptive_LGOCV"}

\item{repeats}{For repeated k-fold cross-validation only: the number of 
complete sets of folds to compute}

\item{number}{Either the number of folds or number of resampling iterations}

\item{...}{Arguments passed to the classification or regression routine}
}
\value{
A list object
}
\description{
Fit predictive models for PLS-DA
}

