% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{center<-}
\alias{center<-}
\title{center}
\usage{
center(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
center
}
\examples{
para <- new("plsDAPara")
center(para) <- TRUE
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
