% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{checkPvaluePlot}
\alias{checkPvaluePlot}
\title{checkPvaluePlot}
\usage{
checkPvaluePlot(file = NULL, group = NULL, fig = "pvalue.png")
}
\arguments{
\item{group}{the group name}

\item{fig}{the file name of output figure}

\item{f1}{a file contained the quantification result of metaX}
}
\value{
none
}
\description{
Plot pvalue distribution
}
