% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{checkQCPlot}
\alias{checkQCPlot}
\title{checkQCPlot}
\usage{
checkQCPlot(f1, f2 = NULL, fig = "test.png", group = NULL)
}
\arguments{
\item{f1}{a file contained the quantification result of metaX}

\item{f2}{a file contained the metabolite identification result}

\item{fig}{the file name of output figure}

\item{group}{the group name}
}
\value{
none
}
\description{
Plot figure for quantification and identification result
}
