% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\docType{methods}
\name{dataClean}
\alias{dataClean}
\title{dataClean}
\usage{
dataClean(para, valueID = "value", sd.factor = 3, snr = 1, ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{valueID}{The name of the column used}

\item{sd.factor}{The factor used to filter peak based on SD}

\item{snr}{The threshold to filter peak}

\item{...}{Other argument}
}
\value{
A metaXpara object.
}
\description{
dataClean
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
p <- dataClean(para)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
