% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{dir.ctrl<-}
\alias{dir.ctrl<-}
\title{dir.ctrl}
\usage{
dir.ctrl(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
dir.ctrl
}
\examples{
para <- new("metaXpara")
dir.ctrl(para) <- "./"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

