% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{filterPeaks}
\alias{filterPeaks}
\title{filterPeaks}
\usage{
filterPeaks(para, ratio = 0.8, omit.negative = TRUE, ...)
}
\arguments{
\item{para}{An object of metaXpara}

\item{ratio}{filter peaks which have missing value more than percent of 
"ratio", default is 0.8}

\item{omit.negative}{A logical value indicates whether omit the negative 
value}

\item{...}{Additional parameters}
}
\value{
An object of metaXpara
}
\description{
filter peaks according to the non-QC sample
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
p <- filterPeaks(para,ratio=0.2)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

