% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{filterQCPeaks}
\alias{filterQCPeaks}
\title{filterQCPeaks}
\usage{
filterQCPeaks(para, ratio = 0.5, omit.negative = TRUE, ...)
}
\arguments{
\item{para}{An object of metaXpara}

\item{ratio}{filter peaks which have missing value more than percent of 
"ratio", default is 0.5}

\item{omit.negative}{A logical value indicates whether omit the negative 
value}

\item{...}{Additional parameters}
}
\value{
An object of metaXpara
}
\description{
filter peaks according to the QC sample
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
p <- filterQCPeaks(para,ratio=0.5)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
