% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{filterQCPeaksByCV}
\alias{filterQCPeaksByCV}
\title{Filter peaks according to the RSD of peaks in QC samples}
\usage{
filterQCPeaksByCV(para, cvFilter = 0.3, valueID = "value", ...)
}
\arguments{
\item{para}{An object of metaXpara}

\item{cvFilter}{Filter peaks with the RSD in QC samples > cvFilter.}

\item{valueID}{The name of the column which will be used.}

\item{...}{Additional parameter}
}
\value{
An object of metaXpara
}
\description{
Filter peaks according to the RSD of peaks in QC samples. 
Usually used after missing value imputation.
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
p <- filterQCPeaksByCV(para)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
