% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{getTable}
\alias{getTable}
\title{Get a table from a metaXpara object.}
\usage{
getTable(para, valueID = "valueNorm", outfile = NULL,
  zero2NA = FALSE)
}
\arguments{
\item{para}{A metaXpara object}

\item{valueID}{The ID of value to extract}
}
\value{
A data.frame
}
\description{
Get a table from a metaXpara object.
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
