% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{group.bw<-}
\alias{group.bw<-}
\title{group.bw}
\usage{
group.bw(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
group.bw
}
\examples{
library(faahKO)
para <- new("metaXpara")
group.bw(para) <- 5
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

