% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{group.bw0<-}
\alias{group.bw0<-}
\title{group.bw0}
\usage{
group.bw0(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
group.bw0
}
\examples{
library(faahKO)
para <- new("metaXpara")
group.bw0(para) <- 10
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
