% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{group.max<-}
\alias{group.max<-}
\title{group.max}
\usage{
group.max(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
group.max
}
\examples{
library(faahKO)
para <- new("metaXpara")
group.max(para) <- 1000
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
