% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{group.minfrac<-}
\alias{group.minfrac<-}
\title{group.minfrac}
\usage{
group.minfrac(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
group.minfrac
}
\examples{
library(faahKO)
para <- new("metaXpara")
group.minfrac(para) <- 0.3
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
