% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{group.minsamp<-}
\alias{group.minsamp<-}
\title{group.minsamp}
\usage{
group.minsamp(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
group.minsamp
}
\examples{
library(faahKO)
para <- new("metaXpara")
group.minsamp(para) <- 1
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
