% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{group.mzwid0<-}
\alias{group.mzwid0<-}
\title{group.mzwid0}
\usage{
group.mzwid0(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
group.mzwid0
}
\examples{
library(faahKO)
para <- new("metaXpara")
group.mzwid0(para) <- 0.015
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

