% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{idres<-}
\alias{idres<-}
\title{idres}
\usage{
idres(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
idres
}
\examples{
library(faahKO)
para <- new("metaXpara")
idres(para) <- data.frame()
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

