% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\name{importDataFromMetaboAnalyst}
\alias{importDataFromMetaboAnalyst}
\title{importDataFromMetaboAnalyst}
\usage{
importDataFromMetaboAnalyst(para, file)
}
\arguments{
\item{para}{an object of metaXpara}

\item{file}{a csv file exported from MetaboAnalyst, which contains peak 
intensity data}
}
\value{
an object of metaXpara
}
\description{
Import peak data from MetaboAnalyst
}

