% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\name{importDataFromQI}
\alias{importDataFromQI}
\title{importDataFromQI}
\usage{
importDataFromQI(para, file, mode = 1, fw = NULL, rt = NULL)
}
\arguments{
\item{para}{an object of metaXpara}

\item{file}{a csv file exported from Progenesis QI, which contains peak 
intensity data}

\item{mode}{1, read the normalized data; 2, read the raw data}

\item{fw}{valid peak width range, for example, it can be set as c(1,30).
The unit is second.}

\item{rt}{valid retention time range, for example, it can be set as c(0.5,9).
The unit is minute}
}
\value{
an object of metaXpara
}
\description{
Import peak data from Progenesis QI.
}
