% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\name{importDataFromXCMS}
\alias{importDataFromXCMS}
\title{importDataFromXCMS}
\usage{
importDataFromXCMS(para, file)
}
\arguments{
\item{para}{an object of metaXpara}

\item{file}{a csv or txt format file exported from XCMS, which contains peak 
intensity data}
}
\value{
an object of metaXpara
}
\description{
Import peak data from XCMS
}

