% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{kfold<-}
\alias{kfold<-}
\title{kfold}
\usage{
kfold(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
kfold
}
\examples{
para <- new("plsDAPara")
kfold(para) <- 5
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

