% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{makeDirectory}
\alias{makeDirectory}
\title{Create directory}
\usage{
makeDirectory(para)
}
\arguments{
\item{para}{A metaXpara object}
}
\value{
none
}
\description{
Create directory
}
\examples{
para <- new("metaXpara")
outdir(para) <- "outdir"
makeDirectory(para)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
