% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{makeMetaboAnalystInput}
\alias{makeMetaboAnalystInput}
\title{Export a csv file which can be used for MetaboAnalyst}
\usage{
makeMetaboAnalystInput(para, rmQC = TRUE, valueID = "valueNorm",
  zero2NA = TRUE, prefix = NA, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{rmQC}{A logical indicates whether remove the QC data}

\item{valueID}{The name of the column which will be used}

\item{zero2NA}{A logical indicates whether convert the value <=0 to NA}

\item{prefix}{The prefix of output file}

\item{...}{Additional parameter}
}
\value{
none
}
\description{
Export a csv file which can be used for MetaboAnalyst
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
makeMetaboAnalystInput(para,valueID="value")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

