% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboliteIdentify.R
\name{metaboliteAnnotation}
\alias{metaboliteAnnotation}
\title{Metabolite identification}
\usage{
metaboliteAnnotation(para, db, delta = 10, mode = 1, unit = "ppm",
  dbType = 1)
}
\arguments{
\item{para}{An object of \code{metaXpara},use the information from the 
rawPeaks to do the identification}

\item{db}{The file name of database}

\item{delta}{The delta}

\item{mode}{The mode of data, positive or negative,1:positive, 2:negative}

\item{unit}{The unit of the delta,1:ppm,2:da. Default is ppm}

\item{dbType}{1=HMDB,2=PubChem,3=KEGG,4=MassBank,5=LipidMaps,6=PlantCyc}
}
\value{
The name of output file
}
\description{
Metabolite identification
}

