% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{class}
\name{metaXpara-class}
\alias{metaXpara-class}
\title{An S4 class to represent the parameters and data for data processing}
\arguments{
\item{para}{A metaXpara object}

\item{value}{New value}
}
\value{
A object of metaXpara
}
\description{
An S4 class to represent the parameters and data for data processing
}
\section{Slots}{

\describe{
\item{\code{dir.case}}{The path names of the NetCDF/mzXML files to read}

\item{\code{dir.ctrl}}{The path names of the NetCDF/mzXML files to read}

\item{\code{sampleListFile}}{The file name of containing the experiment design}

\item{\code{sampleList}}{A data.frame containing the experiment design}

\item{\code{ratioPairs}}{A character containing the ratio pairs, such as "A:B;A:C"}

\item{\code{pairTest}}{Paired test, default is FALSE}

\item{\code{missValueImputeMethod}}{A character of missing value imputation method}

\item{\code{sampleListHead}}{The name of head of \code{sampleListFile}}

\item{\code{outdir}}{The output directory}

\item{\code{prefix}}{The prefix of output file}

\item{\code{xcmsPeakListFile}}{The file of output from \pkg{XCMS}}

\item{\code{fig}}{A \code{list} of file names of figures}

\item{\code{peaksData}}{A \code{data.frame} containing the peaks data}

\item{\code{VIP}}{A \code{data.frame} containing the VIP}

\item{\code{rawPeaks}}{A \code{data.frame} containg the raw peaks data}

\item{\code{xcmsSetObj}}{An object of \code{\link{xcmsSet}}}

\item{\code{quant}}{A \code{data.frame} containing the quantification result}

\item{\code{idres}}{A \code{data.frame} containing the identification result}

\item{\code{xcmsSet.method}}{Method to use for peak detection. See details 
\code{\link{findPeaks}} in package \pkg{XCMS}}

\item{\code{xcmsSet.ppm}}{The maxmial tolerated m/z deviation in consecutive scans, 
in ppm (parts per million)}

\item{\code{xcmsSet.peakwidth}}{Chromatographic peak width, given as 
range (min,max) in seconds}

\item{\code{xcmsSet.snthresh}}{The signal to noise ratio cutoff, definition see 
\code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.prefilter}}{prefilter=c(k,I), 
see \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.mzCenterFun}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.integrate}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.mzdiff}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.noise}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.verbose.columns}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.polarity}}{Filter raw data for positive/negative scans. 
See \code{\link{xcmsSet}}}

\item{\code{xcmsSet.profparam}}{Parameters to use for profile generation. 
See \code{\link{xcmsSet}}}

\item{\code{xcmsSet.nSlaves}}{The number of slaves/cores to be used for parallel 
peak detection. See \code{\link{xcmsSet}}}

\item{\code{xcmsSet.fitgauss}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.sleep}}{The number of seconds to pause between plotting 
peak finding cycles. See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.fwhm}}{See \code{\link{findPeaks.matchedFilter}}}

\item{\code{xcmsSet.max}}{See \code{\link{findPeaks.matchedFilter}}}

\item{\code{xcmsSet.step}}{See \code{\link{findPeaks.matchedFilter}}}

\item{\code{group.bw0}}{See \code{\link{group.density}}}

\item{\code{group.mzwid0}}{See \code{\link{group.density}}}

\item{\code{group.bw}}{See \code{\link{group.density}}}

\item{\code{group.mzwid}}{See \code{\link{group.density}}}

\item{\code{group.minfrac}}{See \code{\link{group.density}}}

\item{\code{group.minsamp}}{See \code{\link{group.density}}}

\item{\code{group.max}}{See \code{\link{group.density}}}

\item{\code{group.sleep}}{See \code{\link{group.density}}}

\item{\code{retcor.method}}{See \code{\link{retcor}}}

\item{\code{retcor.profStep}}{See \code{\link{retcor.obiwarp}}}

\item{\code{retcor.plottype}}{See \code{\link{retcor.obiwarp}}}

\item{\code{qcRlscSpan}}{The value of span for QC-RLSC}
}}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

