% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\docType{methods}
\name{metaXpipe}
\alias{metaXpipe}
\title{metaXpipe}
\usage{
metaXpipe(para, plsdaPara, cvFilter = 0.3, remveOutlier = TRUE,
  outTol = 1.2, doQA = TRUE, doROC = TRUE, qcsc = 0,
  nor.method = "pqn", pclean = TRUE, t = 1, scale = "uv",
  idres = NULL, nor.order = 1, out.rmqc = FALSE, saveRds = TRUE,
  cpu = 0, missValueRatioQC = 0.5, missValueRatioSample = 0.8,
  pcaLabel = "none", ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{plsdaPara}{A plsDAPara object.}

\item{cvFilter}{Filter peaks which cv > cvFilter in QC samples.}

\item{remveOutlier}{Remove outlier samples.}

\item{outTol}{The threshold to remove outlier samples.}

\item{doQA}{Boolean, setting the argument to TRUE will perform plot 
quality figures.}

\item{doROC}{A logical indicates whether to calculate the ROC}

\item{qcsc}{QC-based batch correction, 0=none,1=QC-RLSC(Quality 
control-robust loess signal correction),2=SVR(SVR normalization).}

\item{nor.method}{Normalization method.}

\item{pclean}{Boolean, setting the argument to TRUE to perform 
data cleaning}

\item{t}{Data transformation method. See \code{\link{transformation}}.}

\item{scale}{Data scaling method.}

\item{idres}{A file containing the metabolite identification result}

\item{nor.order}{The order of normalization, only valid when \code{qcsc} is 
TRUE. 1: before QC-based batch correction, 2: after QC-based batch correction.}

\item{out.rmqc}{Boolean, setting the argument to TRUE to remove the QC 
samples for the csv file.}

\item{saveRds}{Boolean, setting the argument to TRUE to save some objects to
disk for debug. Only useful for developer. Default is TRUE.}

\item{cpu}{The number of cpu used, default is all available cpus.}

\item{missValueRatioQC}{The cutoff of the ratio of miss value for the 
features in QC, default is 0.5.}

\item{missValueRatioSample}{The cutoff of the ratio of miss value for the 
features in sample, default is 0.8.}

\item{pcaLabel}{The label used for PCA score plot, "none","order" and 
"sample" are supported. Default is "none"}

\item{...}{Other argument}
}
\value{
A metaXpara object.
}
\description{
metaXpipe
}
\examples{
\dontrun{
## example 1: no QC sample
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
ratioPairs(para) <- "KO:WT"
outdir(para) <- "test"
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
plsdaPara <- new("plsDAPara")
p <- metaXpipe(para,plsdaPara=plsdaPara)

## example 2: has QC samples
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
ratioPairs(para) <- "S:C"
plsdaPara <- new("plsDAPara")
p <- metaXpipe(para,plsdaPara=plsdaPara)
}
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

