% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{method<-}
\alias{method<-}
\title{method}
\usage{
method(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
method
}
\examples{
para <- new("plsDAPara")
method(para) <- "oscorespls"
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
