% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{missValueImputeMethod<-}
\alias{missValueImputeMethod<-}
\title{missValueImputeMethod}
\usage{
missValueImputeMethod(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
missValueImputeMethod
}
\examples{
para <- new("metaXpara")
missValueImputeMethod(para) <- "knn"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

