% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{myCalcAUROC}
\alias{myCalcAUROC}
\title{Classical univariate ROC analysis}
\usage{
myCalcAUROC(para, cgroup, cpu = 0, plot = FALSE, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{cgroup}{Samples used}

\item{cpu}{The number of CPU used}

\item{plot}{A logical indicates whether plot}

\item{...}{Additional parameter}
}
\value{
A metaXpara object
}
\description{
Classical univariate ROC analysis
}
\examples{
\dontrun{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
addValueNorm(para) <- para
res <- myCalcAUROC(para,cgroup=c("S","C"))
}
}
