% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{myPLSDA}
\alias{myPLSDA}
\title{Perform PLS-DA analysis}
\usage{
myPLSDA(x, y, save, select, ...)
}
\arguments{
\item{x}{A matrix of observations}

\item{y}{a vector or matrix of responses}

\item{save}{A logical indicates whether save the \code{plsr} result}

\item{select}{A logical indicates whether select the best component}

\item{...}{Additional parameters}

\item{ncomp}{The number of component used for PLS-DA}

\item{validation}{See \code{\link{plsr}}}

\item{method}{See \code{\link{plsr}}}

\item{k}{k-fold}
}
\value{
The PLS-DA result
}
\description{
Perform PLS-DA analysis
}
\examples{
x <- matrix(rnorm(1000),nrow = 10,ncol = 100)
y <- rep(0:1,5)
res <- myPLSDA(x,y,save=TRUE,ncomp=2,validation="CV",k=7,
    method="oscorespls")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

