% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{ncomp<-}
\alias{ncomp<-}
\title{ncomp}
\usage{
ncomp(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
ncomp
}
\examples{
para <- new("plsDAPara")
ncomp(para) <- 5
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
