% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{normalize}
\alias{normalize}
\title{Normalisation of peak intensity}
\usage{
normalize(para, method = "sum", valueID = "value", norFactor = NULL,
  useClass = FALSE, ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{method}{The normalization method: sum, vsn, quantiles, 
quantiles.robust, sum, pqn, combat and tmm. Default is sum.}

\item{valueID}{The name of the column which will be normalized.}

\item{norFactor}{The factor that will be used for normalization. This is 
usually used in urine data normalization. The factor is the column name in 
sample list file that contains the normalization factor value, such as the 
value "osmolality". Default the value is NULL. Only if the value of 
norFactor is not NULL and the parameter "method" is NULL, this 
normalization will work.}

\item{useClass}{Whether or not to use class information when perform 
batch correction using ComBat method. Default is False.}

\item{...}{Additional parameter}
}
\value{
A metaXpara object.
}
\description{
The normalize method performs normalisation on peak 
intensities.
}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
para <- reSetPeaksData(para)
para <- metaX::normalize(para)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
