% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{nperm<-}
\alias{nperm<-}
\title{nperm}
\usage{
nperm(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
nperm
}
\examples{
para <- new("plsDAPara")
nperm(para) <- 1000
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

