% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{class}
\name{oplsDAPara-class}
\alias{oplsDAPara-class}
\title{An S4 class to represent the parameters for OPLS-DA analysis}
\arguments{
\item{para}{A oplsDAPara object}

\item{value}{New value}
}
\value{
A object of oplsDAPara
}
\description{
An S4 class to represent the parameters for OPLS-DA analysis
}
\section{Slots}{

\describe{
\item{\code{scale}}{The method used to scale the data, 
see \code{\link{preProcess}} in \pkg{metaX}}

\item{\code{center}}{A logical which indicates if the matrix should be mean 
centred or not}

\item{\code{t}}{The method used to transform the data, 
see \code{\link{transformation}} in \pkg{metaX}}

\item{\code{ncomp}}{The number of components used for OPLS-DA, default is 2}

\item{\code{northo}}{The number of orthogonal components}

\item{\code{nperm}}{The number of permutations, default is 200}

\item{\code{kfold}}{The number of folds for cross-validation, default is 7}

\item{\code{do}}{A logical which indicates whether to do the plsDA analysis, 
default is TRUE}
}}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

