% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{outdir<-}
\alias{outdir<-}
\title{outdir}
\usage{
outdir(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
outdir
}
\examples{
para <- new("metaXpara")
outdir(para) <- "outdir"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

