% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnalysis.R
\name{pathwayAnalysis}
\alias{pathwayAnalysis}
\title{Pathway analysis}
\usage{
pathwayAnalysis(id, id.type = "hmdb", outfile)
}
\arguments{
\item{id}{A vector of metabolite IDs}

\item{id.type}{The type of metabolite ID type, default is hmdb.}

\item{outfile}{The output file name}
}
\value{
A data.frame object
}
\description{
Pathway analysis
}
\examples{
\dontrun{
res <- pathwayAnalysis(id=c("HMDB00060","HMDB00056","HMDB00064"),
    outfile="pathway.csv")
head(res)
}
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
