% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{peakFinder}
\alias{peakFinder}
\title{Peak detection by using XCMS package}
\usage{
peakFinder(para, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{...}{Additional parameter}
}
\value{
A metaXpara object
}
\description{
peakFinder takes a set of MS sample data and performs a peak 
detection, retention time correction and peak grouping steps using XCMS 
package.
}
\examples{
\dontrun{
library(faahKO)
para <- new("metaXpara")
dir.case(para) <- system.file("cdf/KO", package = "faahKO")
dir.ctrl(para) <- system.file("cdf/WT", package = "faahKO")
## set parameters for peak picking
xcmsSet.peakwidth(para) <- c(20,50)
xcmsSet.snthresh(para) <- 10
xcmsSet.prefilter(para) <- c(3,100)
xcmsSet.noise(para) <- 0
xcmsSet.nSlaves(para) <- 4
## run peak picking
p <- peakFinder(para)
}
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
