% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{peakStat}
\alias{peakStat}
\title{Do the univariate and multivariate statistical analysis}
\usage{
peakStat(para, plsdaPara, doROC = TRUE, saveRds = TRUE, ...)
}
\arguments{
\item{para}{A \code{metaXpara} object}

\item{plsdaPara}{A \code{plsDAPara} object}

\item{doROC}{A logical indicates whether to calculate the ROC}

\item{saveRds}{Boolean, setting the argument to TRUE to save some objects to
disk for debug. Only useful for developer. Default is TRUE.}

\item{...}{Additional parameter}

\item{pcaLabel}{The label used for PCA score plot, "none","order" and 
"sample" are supported. Default is "none"}
}
\value{
none

An object of \code{metaXpara}
}
\description{
Do the univariate and multivariate statistical analysis
}
\examples{
\dontrun{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
ratioPairs(para) <- "S:C"
addValueNorm(para) <- para
plsdaPara <- new("plsDAPara")
res <- peakStat(para,plsdaPara)
}
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
