% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{plotCorHeatmap}
\alias{plotCorHeatmap}
\title{Plot correlation heatmap}
\usage{
plotCorHeatmap(para, valueID = "value", samples = NA, label = "order",
  width = 6, cor.method = "spearman", height = 6, anno = FALSE,
  cluster = FALSE, shownames = FALSE, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{valueID}{The name of the column that used for plot}

\item{samples}{Samples used for plot}

\item{label}{Label to show in figure}

\item{width}{The width of the graphics region in inches. 
The default values are 6.}

\item{cor.method}{Method used for correlation}

\item{height}{The height of the graphics region in inches. 
The default values are 6.}

\item{anno}{A logical value indicates whether to plot heatmap with 
annotating class information}

\item{cluster}{A logical value indicates whether to do the cluster when 
anno is TRUE}

\item{shownames}{A logical indicates whether show names when plot}

\item{...}{Additional parameter}
}
\value{
The fig name
}
\description{
This function plots correlation heatmap.
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
plotCorHeatmap(para,valueID="value",samples=NULL,width=6,anno=TRUE)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

