% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotCV}
\alias{plotCV}
\title{Plot the CV distribution of peaks in each group}
\usage{
plotCV(x, valueID = "value", ...)
}
\arguments{
\item{x}{A metaXpara object}

\item{valueID}{The name of the column that used as peak intensity}

\item{...}{Additional parameter}
}
\value{
none
}
\description{
Plot the CV distribution of peaks in each group.
}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
para <- reSetPeaksData(para)
plotCV(para)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
