% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotHeatMap}
\alias{plotHeatMap}
\title{Plot heatmap}
\usage{
plotHeatMap(para, valueID = "valueNorm", log = TRUE, rmQC = TRUE,
  zero2na = FALSE, colors = "none", anno = NULL, width = 12,
  height = 8, saveRds = TRUE, show_rownames = FALSE,
  show_colnames = FALSE, classCol = NULL, seriation = FALSE, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{valueID}{The name of the column that used for plot}

\item{log}{A logical indicating whether to log the data}

\item{rmQC}{A logical indicating whether to remove the QC samples}

\item{zero2na}{A logical indicating whether to convert the value <=0 to NA}

\item{colors}{Color for heatmap}

\item{anno}{A file or a data frame including sample annotation data.}

\item{width}{The width of the graphics region in inches. 
The default values are 12.}

\item{height}{The height of the graphics region in inches. 
The default values are 8.}

\item{saveRds}{Boolean, setting the argument to TRUE to save some objects to
disk for debug. Only useful for developer. Default is TRUE.}

\item{...}{Additional parameter}
}
\value{
none
}
\description{
This function plots heatmap.
}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
plotHeatMap(para,valueID="value",width=6)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
