% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotIntDistr}
\alias{plotIntDistr}
\title{Plot the distribution of the peaks intensity}
\usage{
plotIntDistr(x, width = 14, sortBy = "order", rmNA = TRUE,
  ylim = NULL, ...)
}
\arguments{
\item{x}{A metaXpara object.}

\item{width}{The width of pdf, default is 14.}

\item{sortBy}{Sort the samples according to order (default), batch or class}

\item{rmNA}{Whether or not to consider NA, default is TRUE. When this value is FALSE,
NA will be considered as 0 then all values are +1.}

\item{...}{Additional parameter}
}
\value{
The figure name
}
\description{
Plot the distribution of the peaks intensity for both raw 
intensity and normalized intensity.
}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
para <- reSetPeaksData(para)
plotIntDistr(para)
## after normalization
para <- metaX::normalize(para)
plotIntDistr(para)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
