% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotLoading}
\alias{plotLoading}
\title{Plot figures for PCA/PLS-DA loadings}
\usage{
plotLoading(object, out.tol = 0.9, show.outlier = FALSE, label = 0,
  fig = "loading.png")
}
\arguments{
\item{object}{object of pcaRes or PLS-DA}

\item{out.tol}{control the points to show labels}

\item{show.outlier}{Logical. The points will be colour-coded when the value is set as TRUE.
Default is FALSE}

\item{label}{0=>only show part of the labels, 1=>show all the labels, 3=none labels}
}
\value{
none
}
\description{
Plot figure for PCA/PLS-DA loadings
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
para <- transformation(para,valueID = "value")
res <- metaX::plotPCA(para,valueID="value",scale="uv",center=TRUE)
plotLoading(res$pca,fig="loading.png")
}
\seealso{
\code{\link{plotPCA}}
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
