% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot correlation network map}
\usage{
plotNetwork(para, group, valueID = "value", cor.thr = 0.95,
  degree.thr = 10, size.factor = 0.5, layout = layout_in_circle,
  showPlot = FALSE, graph_format = "gml", ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{group}{Samples used for plot}

\item{valueID}{The name of the column that used for plot}

\item{cor.thr}{Threshold of correlation}

\item{degree.thr}{Threshold of degree of node}

\item{size.factor}{Node size factor for plot}

\item{layout}{layout for plotting}

\item{showPlot}{Whether or not to print the figure to screen}

\item{graph_format}{The file format for graph data to save. Currently, 
"edgelist", "pajek", "ncol", "lgl", "graphml", "dimacs", "gml", "dot" and 
"leda" are supported.}

\item{...}{Additional parameter}
}
\value{
An object of igraph
}
\description{
Plot correlation network map
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
gg <- plotNetwork(para,group=c("S","C"),degree.thr = 10,cor.thr = 0.8)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

