% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\docType{methods}
\name{plotPeakBox}
\alias{plotPeakBox}
\title{Plot boxplot for each feature}
\usage{
plotPeakBox(para, samples, log = FALSE, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{samples}{Sample class used}

\item{log}{Whether log transform or not}

\item{...}{Additional parameters}
}
\value{
The output figure name.
}
\description{
Plot boxplot for each feature
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)[1:20,]
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
addValueNorm(para) <- para
plotPeakBox(para,samples=c("S","C"))
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

