% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotPeakNumber}
\alias{plotPeakNumber}
\title{Plot the distribution of the peaks number}
\usage{
plotPeakNumber(x, ...)
}
\arguments{
\item{x}{A metaXpara object}

\item{...}{Additional parameter}
}
\value{
The figure name
}
\description{
Plot the distribution of the raw peaks number without 
post-processing. This function not only generates a figure, but also 
saves the information of peaks number into a file.
}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
plotPeakNumber(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

