% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotPLSDA}
\alias{plotPLSDA}
\title{Plot PLS-DA figure}
\usage{
plotPLSDA(para, label = "order", valueID = "valueNorm", ncomp = 5, ...)
}
\arguments{
\item{para}{A \code{metaXpara} object}

\item{label}{The label used for plot PLS-DA figure, default is "order"}

\item{valueID}{The name of the column which will be used}

\item{ncomp}{The number of components used for PLS-DA}

\item{...}{Additional parameter}
}
\value{
none
}
\description{
Plot PLS-DA figure
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
para <- transformation(para,valueID = "value")
para <- preProcess(para=para,scale = "uv",center = TRUE,valueID = "value")
plotPLSDA(para,valueID="value")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

