% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotQC}
\alias{plotQC}
\title{Plot the correlation change of the QC samples.}
\usage{
plotQC(para, valueID = "valueNorm", step = 4, log = TRUE, width = 8,
  height = 4, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{valueID}{The name of the column that used for plot}

\item{step}{The step value of calculate the cor of the samples. Default is 4.}

\item{log}{A logical indicating whether to log the data}

\item{width}{The width of the graphics region in inches. 
The default values are 8.}

\item{height}{The height of the graphics region in inches. 
The default values are 4.}

\item{...}{Additional parameter}
}
\value{
none
}
\description{
Plot the correlation change of the QC samples.
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
plotQC(para,valueID="value")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

