% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotQCRLSC}
\alias{plotQCRLSC}
\title{Plot figures for QC-RLSC}
\usage{
plotQCRLSC(para, maxf = 100)
}
\arguments{
\item{para}{A \code{metaXpara} object}

\item{maxf}{The number of features to plot}
}
\value{
none
}
\description{
Plot figures for QC-RLSC
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)[1:20,]
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
res <- doQCRLSC(para,cpu=1)
plotQCRLSC(res$metaXpara)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
\seealso{
\code{\link{doQCRLSC}}
}

