% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{plotROC}
\alias{plotROC}
\title{Perform biomarker analysis and plot ROC curve}
\usage{
plotROC(para, group, fID, valueID = "valueNorm", method = "rf",
  ratio = 2/3, k = 100, plotCICurve = TRUE, ...)
}
\arguments{
\item{para}{An object of \code{metaXpara}}

\item{group}{The sample class used}

\item{valueID}{The column name used}

\item{method}{Method for modeling}

\item{ratio}{(The number of sample for training)/(The number of sample for 
Testing)}

\item{k}{The number of sampling. Default is 100.}

\item{plotCICurve}{Whether plot confidence interval curve in the ROC figure}

\item{...}{Additional parameters}
}
\value{
The result of feature selection and modeling
}
\description{
Perform biomarker analysis and plot ROC curve
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
para <- transformation(para,valueID = "value")
# library(doMC)
# registerDoMC(cores = 4)
res <- featureSelection(para,group=c("C","S"), plot_roc=FALSE)
plotROC(para,group=c("C","S"),valueID="value",fID=res$optVariables)
}
