% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{plotTIC}
\alias{plotTIC}
\title{Plot chromatogram of total ion count}
\usage{
plotTIC(file, corBatch = FALSE, corOrder = TRUE)
}
\arguments{
\item{corBatch}{Code colour according to batch number}

\item{corOrder}{Code colour according to injection order}
}
\value{
none
}
\description{
Plot chromatogram of total ion count
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
