% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotTreeMap}
\alias{plotTreeMap}
\title{Plot Phylogenies for samples}
\usage{
plotTreeMap(para, valueID = "valueNorm", log = TRUE, rmQC = TRUE,
  nc = 8, treeType = "fan", width = 8, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{valueID}{The name of the column that used for plot}

\item{log}{A logical indicating whether to log the data}

\item{rmQC}{A logical indicating whether to remove the QC samples}

\item{nc}{The number of clusters}

\item{treeType}{A character string specifying the type of phylogeny to be 
drawn; it must be one of "phylogram" (the default), "cladogram", "fan", 
"unrooted", "radial" or any unambiguous abbreviation of these.}

\item{width}{The width and height of the graphics region in inches. 
The default values are 8.}

\item{...}{Additional parameter}
}
\value{
none
}
\description{
This function plots phylogenetic trees for samples.
}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
para <- reSetPeaksData(para)
plotTreeMap(para,valueID="value")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

