% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{class}
\name{plsDAPara-class}
\alias{plsDAPara-class}
\title{An S4 class to represent the parameters for PLS-DA analysis}
\arguments{
\item{para}{A oplsDAPara object}

\item{value}{New value}
}
\value{
A object of plsDAPara
}
\description{
An S4 class to represent the parameters for PLS-DA analysis
}
\section{Slots}{

\describe{
\item{\code{scale}}{The method used to scale the data, 
see \code{\link{preProcess}} in \pkg{metaX}}

\item{\code{center}}{A logical which indicates if the matrix should be mean 
centred or not}

\item{\code{t}}{The method used to transform the data, 
see \code{\link{transformation}} in \pkg{metaX}}

\item{\code{validation}}{The method for validation, default is "CV"}

\item{\code{ncomp}}{The number of components used for PLS-DA, default is 2}

\item{\code{nperm}}{The number of permutations, default is 200}

\item{\code{kfold}}{The number of folds for cross-validation, default is 7}

\item{\code{do}}{A logical which indicates whether to do the plsDA analysis, 
default is TRUE}

\item{\code{method}}{The method used in PLS-DA. See \code{\link{plsr}} 
in \pkg{pls}}

\item{\code{cpu}}{The number of cpus used, default is all cpus.}
}}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

