% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAnalysis.R
\name{powerAnalyst}
\alias{powerAnalyst}
\title{Power Analysis}
\usage{
powerAnalyst(para, group, valueID = "value", log = TRUE, maxInd = 1000,
  fdr = 0.1, showPlot = FALSE)
}
\arguments{
\item{para}{An metaXpara object}

\item{group}{A vector of sample names}

\item{valueID}{The column name used}

\item{log}{A logical indicating whether transform the data with log2}

\item{maxInd}{max sample number}

\item{fdr}{The FDR threshold}

\item{showPlot}{Whether or not to print the figure to screen}
}
\value{
An value
}
\description{
Power Analysis
}
\examples{
\dontrun{ 
library(reshape2)
library(dplyr)
a <- read.csv("http://www.metaboanalyst.ca/MetaboAnalyst/resources/data/power_example.csv")
peaksData <- melt(a,id.vars = c("Diet","Sample"),
    value.name = "value",variable.name = "ID")
peaksData <- dplyr::rename(peaksData,class=Diet,sample=Sample)
para <- new("metaXpara")
peaksData(para) <- peaksData
para <- missingValueImpute(para)
para <- metaX::normalize(para)
para <- transformation(para,valueID = "value")
para <- preProcess(para,scale = "pareto",valueID="value")
powerAnalyst(para,group=c("case","control"),log=FALSE,maxInd=200)
}
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

