% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{prefix<-}
\alias{prefix<-}
\title{prefix}
\usage{
prefix(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
prefix
}
\examples{
para <- new("metaXpara")
prefix(para) <- "test"
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
