% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{qcRlscSpan<-}
\alias{qcRlscSpan<-}
\title{qcRlscSpan}
\usage{
qcRlscSpan(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
qcRlscSpan
}
\examples{
library(faahKO)
para <- new("metaXpara")
qcRlscSpan(para) <- 0.4
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
