% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{ratioPairs<-}
\alias{ratioPairs<-}
\title{ratioPairs}
\usage{
ratioPairs(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
ratioPairs
}
\examples{
para <- new("metaXpara")
ratioPairs(para) <- "1:2"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

