% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{rawPeaks<-}
\alias{rawPeaks<-}
\title{rawPeaks}
\usage{
rawPeaks(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
rawPeaks
}
\examples{
library(faahKO)
para <- new("metaXpara")
rawPeaks(para) <- data.frame()
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
