% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\docType{methods}
\name{removeSample}
\alias{removeSample}
\title{Remove samples from the metaXpara object}
\usage{
removeSample(para, rsamples, ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{rsamples}{The samples needed to be removed}

\item{...}{Other argument}
}
\value{
A metaXpara object.
}
\description{
Remove samples from the metaXpara object
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
new_para <- removeSample(para,rsamples=c("batch01_QC01"))
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

