% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{retcor.method<-}
\alias{retcor.method<-}
\title{retcor.method}
\usage{
retcor.method(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
retcor.method
}
\examples{
library(faahKO)
para <- new("metaXpara")
retcor.method(para) <- "obiwarp"
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
