% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{retcor.plottype<-}
\alias{retcor.plottype<-}
\title{retcor.plottype}
\usage{
retcor.plottype(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
retcor.plottype
}
\examples{
library(faahKO)
para <- new("metaXpara")
retcor.plottype(para) <- "deviation"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

