% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{retcor.profStep<-}
\alias{retcor.profStep<-}
\title{retcor.profStep}
\usage{
retcor.profStep(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
retcor.profStep
}
\examples{
library(faahKO)
para <- new("metaXpara")
retcor.profStep(para) <- 0.005
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
