% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{runOPLSDA}
\alias{runOPLSDA}
\title{run OPLS-DA}
\usage{
runOPLSDA(para, oplsdaPara, auc = TRUE, sample = NULL,
  valueID = "valueNorm", label = "order", labelSize = 4,
  pointSize = 1.4, ...)
}
\arguments{
\item{para}{An object of metaXpara}

\item{oplsdaPara}{An object of oplsDAPara}

\item{sample}{Sample class}

\item{valueID}{The name of column used}

\item{label}{The label used for plot}

\item{labelSize}{The size of label, default is 4}

\item{pointSize}{The size of point, default is 1.4}

\item{...}{additional arguments}
}
\value{
A list object
}
\description{
Perform the OPLS-DA analysis
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
oplsdaPara <- new("oplsDAPara")
runOPLSDA(para,oplsdaPara,sample=c("C","S"),valueID="value",labelSize = 2)
}
