% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{sampleListFile<-}
\alias{sampleListFile<-}
\title{sampleListFile}
\usage{
sampleListFile(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
sampleListFile
}
\examples{
para <- new("metaXpara")
sampleListFile(para) <- "sample.txt"
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
