% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{scale<-}
\alias{scale<-}
\title{scale}
\usage{
scale(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
scale
}
\examples{
para <- new("plsDAPara")
scale(para) <- "uv"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

