% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{selectBestComponent}
\alias{selectBestComponent}
\title{Select the best component for PLS-DA}
\usage{
selectBestComponent(para, np = 10, sample = NULL, t = 1,
  method = "oscorespls", scale = NULL, center = TRUE,
  valueID = "valueNorm", validation = "CV", k = 7, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{np}{The number of max component}

\item{sample}{The sample class used}

\item{t}{Method used to transform the data}

\item{method}{See \code{\link{plsr}}}

\item{scale}{Method used to scale the data}

\item{center}{Centering}

\item{valueID}{The name of column contained the data}

\item{validation}{See \code{\link{plsr}}}

\item{k}{k-fold}

\item{...}{Additional parameter}
}
\value{
A list
}
\description{
Select the best component for PLS-DA
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
selectBestComponent(para,np=10,sample=c("S","C"),scale="uv",valueID="value")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

