% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\name{svrNormalize}
\alias{svrNormalize}
\title{Batch correction using SVR normalization}
\usage{
svrNormalize(para, ntop = 5, impute = TRUE, cpu = 0)
}
\arguments{
\item{para}{A metaXpara object}

\item{ntop}{Default is top 5 correlated peaks}

\item{impute}{Whether do the imputation before and after normalization}

\item{cpu}{The number of cpu used, default is 0 which means using all cpu.}
}
\value{
A new object of metaXpara
}
\description{
Batch correction using SVR normalization
}
\references{
X. Shen, X. Gong, Y. Cai, Y. Guo, J. Tu, H. Li, T.Zhang, 
J. Wang, F. Xue, and Z.-J. Zhu, Normalization and 
Integration of Large-Scale Metabolomics Data Using Support Vector 
Regression, Metabolomics, 2016, 12: 89
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
