% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{t<-}
\alias{t<-}
\title{t}
\usage{
t(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
t
}
\examples{
para <- new("plsDAPara")
t(para) <- 1
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

