% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{validation<-}
\alias{validation<-}
\title{validation}
\usage{
validation(para) <- value
}
\arguments{
\item{para}{An object of plsDAPara}

\item{value}{value}
}
\value{
An object of plsDAPara
}
\description{
validation
}
\examples{
para <- new("plsDAPara")
validation(para) <- "CV"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

