% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.fitgauss<-}
\alias{xcmsSet.fitgauss<-}
\title{xcmsSet.fitgauss}
\usage{
xcmsSet.fitgauss(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.fitgauss
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.fitgauss(para) <- FALSE
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
