% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.fwhm<-}
\alias{xcmsSet.fwhm<-}
\title{xcmsSet.fwhm}
\usage{
xcmsSet.fwhm(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.fwhm
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.fwhm(para) <- 30
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
