% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.integrate<-}
\alias{xcmsSet.integrate<-}
\title{xcmsSet.integrate}
\usage{
xcmsSet.integrate(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.integrate
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.integrate(para) <- 1
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
