% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.max<-}
\alias{xcmsSet.max<-}
\title{xcmsSet.max}
\usage{
xcmsSet.max(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.max
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.max(para) <- 5
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
