% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.method<-}
\alias{xcmsSet.method<-}
\title{xcmsSet.method}
\usage{
xcmsSet.method(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.method
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.method(para) <- "centWave"
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

