% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.mzCenterFun<-}
\alias{xcmsSet.mzCenterFun<-}
\title{xcmsSet.mzCenterFun}
\usage{
xcmsSet.mzCenterFun(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.mzCenterFun
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.mzCenterFun(para) <- "wMean"
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
