% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.mzdiff<-}
\alias{xcmsSet.mzdiff<-}
\title{xcmsSet.mzdiff}
\usage{
xcmsSet.mzdiff(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.mzdiff
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.mzdiff(para) <- -0.001
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

