% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.noise<-}
\alias{xcmsSet.noise<-}
\title{xcmsSet.noise}
\usage{
xcmsSet.noise(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.noise
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.noise(para) <- 1000
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

