% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.nSlaves<-}
\alias{xcmsSet.nSlaves<-}
\title{xcmsSet.nSlaves}
\usage{
xcmsSet.nSlaves(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.nSlaves
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.nSlaves(para) <- 8
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
