% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.peakwidth<-}
\alias{xcmsSet.peakwidth<-}
\title{xcmsSet.peakwidth}
\usage{
xcmsSet.peakwidth(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.peakwidth
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.peakwidth(para) <- 12
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

