% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.polarity<-}
\alias{xcmsSet.polarity<-}
\title{xcmsSet.polarity}
\usage{
xcmsSet.polarity(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.polarity
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.polarity(para) <- "positive"
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
