% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.ppm<-}
\alias{xcmsSet.ppm<-}
\title{xcmsSet.ppm}
\usage{
xcmsSet.ppm(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.ppm
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.ppm(para) <- 10
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
