% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.prefilter<-}
\alias{xcmsSet.prefilter<-}
\title{xcmsSet.prefilter}
\usage{
xcmsSet.prefilter(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.prefilter
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.prefilter(para) <- c(1,5000)
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
