% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.sleep<-}
\alias{xcmsSet.sleep<-}
\title{xcmsSet.sleep}
\usage{
xcmsSet.sleep(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.sleep
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.sleep(para) <- 0
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

