% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.snthresh<-}
\alias{xcmsSet.snthresh<-}
\title{xcmsSet.snthresh}
\usage{
xcmsSet.snthresh(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.snthresh
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.snthresh(para) <- 5
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
