% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.step<-}
\alias{xcmsSet.step<-}
\title{xcmsSet.step}
\usage{
xcmsSet.step(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.step
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.step(para) <- 0.1
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

