% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSet.verbose.columns<-}
\alias{xcmsSet.verbose.columns<-}
\title{xcmsSet.verbose.columns}
\usage{
xcmsSet.verbose.columns(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSet.verbose.columns
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSet.verbose.columns(para) <- FALSE
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

