% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{xcmsSetObj<-}
\alias{xcmsSetObj<-}
\title{xcmsSetObj}
\usage{
xcmsSetObj(para) <- value
}
\arguments{
\item{para}{An object of metaXpara}

\item{value}{value}
}
\value{
An object of metaXpara
}
\description{
xcmsSetObj
}
\examples{
library(faahKO)
para <- new("metaXpara")
xcmsSetObj(para) <- faahko
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

