% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{zero2NA}
\alias{zero2NA}
\title{Convert the value <=0 to NA}
\usage{
zero2NA(x, valueID = "value", ...)
}
\arguments{
\item{x}{An object of data}

\item{valueID}{The name of the column which will be used}

\item{...}{Additional parameters}
}
\value{
An object of \code{metaXpara}
}
\description{
Convert the value <=0 to NA
}
\examples{
\donttest{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- zero2NA(para)
}
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

