% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tool.R
\docType{methods}
\name{addIdentInfo}
\alias{addIdentInfo}
\alias{addIdentInfo,metaXpara,character-method}
\title{Add identification result into metaXpara object}
\usage{
addIdentInfo(para, file, ...)

\S4method{addIdentInfo}{metaXpara,character}(para, file, ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{file}{The file name which contains the identification result}

\item{...}{Other argument}
}
\value{
A metaXpara object.
}
\description{
Add identification result into metaXpara object
}
\section{Methods (by class)}{
\itemize{
\item \code{para = metaXpara,file = character}: 
}}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

