% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{calcVIP}
\alias{calcVIP}
\alias{calcVIP,ANY-method}
\title{Calculate the VIP for PLS-DA}
\usage{
calcVIP(x, ncomp, ...)

\S4method{calcVIP}{ANY}(x, ncomp, ...)
}
\arguments{
\item{x}{An object of output from \code{plsr}}

\item{ncomp}{The number of component used in PLS-DA}

\item{...}{Additional parameters}
}
\value{
An vector of VIP value
}
\description{
Calculate the VIP for PLS-DA
}
\section{Methods (by class)}{
\itemize{
\item \code{ANY}: 
}}
\examples{
library(pls)
x <- matrix(rnorm(1000),nrow = 10,ncol = 100)
y <- rep(0:1,5)
res <- plsr(y~x)
calcVIP(res,2)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

