% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network.R
\name{cor.network}
\alias{cor.network}
\title{Correlation network analysis}
\usage{
cor.network(para, group, valueID = "value", cor.method = "spearman",
  threshold = 0.1, p.adjust.methods = "BH")
}
\arguments{
\item{para}{A metaXpara object}

\item{group}{Samples used for plot}

\item{valueID}{The name of the column that used for plot}

\item{cor.method}{Method for correlation:"pearson","spearman" or "kendall"}

\item{threshold}{A threshold of significance levels of
differential correlation}

\item{p.adjust.methods}{c("local", holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none")}

\item{...}{Additional parameter}
}
\value{
The name of result file
}
\description{
Correlation network analysis
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
resfile <- cor.network(para,group=c("S","C"))
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

