% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fselection.R
\name{featureSelection}
\alias{featureSelection}
\title{Feature selection and modeling}
\usage{
featureSelection(para, group, method = "rf", valueID = "value", fold = 5,
  repeats = 10, verbose = FALSE, ...)
}
\arguments{
\item{para}{An object of \code{metaXpara}}

\item{group}{The sample class used}

\item{method}{Method for feature selection and modeling}

\item{valueID}{The column name used}

\item{fold}{k-fold}

\item{repeats}{The repeat number}

\item{verbose}{Whether output or not}

\item{...}{Additional parameters}
}
\value{
The result of feature selection and modeling
}
\description{
Feature selection and modeling
}

