% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{getPeaksTable}
\alias{getPeaksTable}
\alias{getPeaksTable,metaXpara-method}
\title{Get a data.frame which contained the peaksData in metaXpara}
\usage{
getPeaksTable(para, sample = NULL, valueID = "value")

\S4method{getPeaksTable}{metaXpara}(para, sample = NULL, valueID = "value")
}
\arguments{
\item{para}{An object of data}

\item{sample}{Sample class used}

\item{valueID}{The column name used}
}
\value{
A data.frame
}
\description{
Get a data.frame which contained the peaksData in metaXpara
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 
}}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
res <- getPeaksTable(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

