% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{hasQC}
\alias{hasQC}
\alias{hasQC,metaXpara-method}
\title{Judge whether the data has QC samples}
\usage{
hasQC(para, ...)

\S4method{hasQC}{metaXpara}(para, ...)
}
\arguments{
\item{para}{An object of data}

\item{...}{Additional parameters}
}
\value{
A logical value
}
\description{
Judge whether the data has QC samples
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 
}}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
hasQC(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

