% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{makeDirectory}
\alias{makeDirectory}
\alias{makeDirectory,metaXpara-method}
\title{Create directory}
\usage{
makeDirectory(para)

\S4method{makeDirectory}{metaXpara}(para)
}
\arguments{
\item{para}{A metaXpara object}
}
\value{
none
}
\description{
Create directory
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 
}}
\examples{
para <- new("metaXpara")
outdir(para) <- "outdir"
makeDirectory(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

