% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MetaboliteIdentify.R
\name{metaboliteAnnotation}
\alias{metaboliteAnnotation}
\title{Metabolite identification}
\usage{
metaboliteAnnotation(para, db, delta, mode, unit)
}
\arguments{
\item{para}{An object of \code{metaXpara}}

\item{db}{The file name of database}

\item{delta}{The delta}

\item{mode}{The mode of data, positive or negative}

\item{unit}{The unit of the delta}
}
\value{
The name of output file
}
\description{
Metabolite identification
}

