% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{class}
\name{metaXpara-class}
\alias{addValueNorm<-,metaXpara,metaXpara-method}
\alias{dir.case<-,metaXpara-method}
\alias{dir.ctrl<-,metaXpara-method}
\alias{group.bw0<-,metaXpara-method}
\alias{group.bw<-,metaXpara-method}
\alias{group.max<-,metaXpara-method}
\alias{group.minfrac<-,metaXpara-method}
\alias{group.minsamp<-,metaXpara-method}
\alias{group.mzwid0<-,metaXpara-method}
\alias{group.mzwid<-,metaXpara-method}
\alias{group.sleep<-,metaXpara-method}
\alias{idres<-,metaXpara-method}
\alias{metaXpara-class}
\alias{missValueImputeMethod<-,metaXpara-method}
\alias{outdir<-,metaXpara-method}
\alias{peaksData<-,metaXpara-method}
\alias{prefix<-,metaXpara-method}
\alias{qcRlscSpan<-,metaXpara-method}
\alias{ratioPairs<-,metaXpara-method}
\alias{rawPeaks<-,metaXpara-method}
\alias{retcor.method<-,metaXpara-method}
\alias{retcor.plottype<-,metaXpara-method}
\alias{retcor.profStep<-,metaXpara-method}
\alias{sampleListFile<-,metaXpara-method}
\alias{xcmsSet.fitgauss<-,metaXpara-method}
\alias{xcmsSet.fwhm<-,metaXpara-method}
\alias{xcmsSet.integrate<-,metaXpara-method}
\alias{xcmsSet.max<-,metaXpara-method}
\alias{xcmsSet.method<-,metaXpara-method}
\alias{xcmsSet.mzCenterFun<-,metaXpara-method}
\alias{xcmsSet.mzdiff<-,metaXpara-method}
\alias{xcmsSet.nSlaves<-,metaXpara-method}
\alias{xcmsSet.noise<-,metaXpara-method}
\alias{xcmsSet.peakwidth<-,metaXpara-method}
\alias{xcmsSet.polarity<-,metaXpara-method}
\alias{xcmsSet.ppm<-,metaXpara-method}
\alias{xcmsSet.prefilter<-,metaXpara-method}
\alias{xcmsSet.profparam<-,metaXpara-method}
\alias{xcmsSet.sleep<-,metaXpara-method}
\alias{xcmsSet.snthresh<-,metaXpara-method}
\alias{xcmsSet.step<-,metaXpara-method}
\alias{xcmsSet.verbose.columns<-,metaXpara-method}
\alias{xcmsSetObj<-,metaXpara-method}
\title{An S4 class to represent the parameters and data for data processing}
\usage{
\S4method{dir.case}{metaXpara}(para) <- value

\S4method{dir.ctrl}{metaXpara}(para) <- value

\S4method{sampleListFile}{metaXpara}(para) <- value

\S4method{ratioPairs}{metaXpara}(para) <- value

\S4method{missValueImputeMethod}{metaXpara}(para) <- value

\S4method{outdir}{metaXpara}(para) <- value

\S4method{prefix}{metaXpara}(para) <- value

\S4method{peaksData}{metaXpara}(para) <- value

\S4method{addValueNorm}{metaXpara,metaXpara}(para) <- value

\S4method{rawPeaks}{metaXpara}(para) <- value

\S4method{idres}{metaXpara}(para) <- value

\S4method{xcmsSet.method}{metaXpara}(para) <- value

\S4method{xcmsSet.ppm}{metaXpara}(para) <- value

\S4method{xcmsSet.peakwidth}{metaXpara}(para) <- value

\S4method{xcmsSet.snthresh}{metaXpara}(para) <- value

\S4method{xcmsSet.prefilter}{metaXpara}(para) <- value

\S4method{xcmsSet.mzCenterFun}{metaXpara}(para) <- value

\S4method{xcmsSet.integrate}{metaXpara}(para) <- value

\S4method{xcmsSet.mzdiff}{metaXpara}(para) <- value

\S4method{xcmsSet.noise}{metaXpara}(para) <- value

\S4method{xcmsSet.verbose.columns}{metaXpara}(para) <- value

\S4method{xcmsSet.polarity}{metaXpara}(para) <- value

\S4method{xcmsSet.profparam}{metaXpara}(para) <- value

\S4method{xcmsSet.nSlaves}{metaXpara}(para) <- value

\S4method{xcmsSet.fitgauss}{metaXpara}(para) <- value

\S4method{xcmsSet.sleep}{metaXpara}(para) <- value

\S4method{xcmsSet.fwhm}{metaXpara}(para) <- value

\S4method{xcmsSet.max}{metaXpara}(para) <- value

\S4method{xcmsSet.step}{metaXpara}(para) <- value

\S4method{group.bw0}{metaXpara}(para) <- value

\S4method{group.mzwid0}{metaXpara}(para) <- value

\S4method{group.bw}{metaXpara}(para) <- value

\S4method{group.mzwid}{metaXpara}(para) <- value

\S4method{group.minfrac}{metaXpara}(para) <- value

\S4method{group.minsamp}{metaXpara}(para) <- value

\S4method{group.max}{metaXpara}(para) <- value

\S4method{group.sleep}{metaXpara}(para) <- value

\S4method{retcor.method}{metaXpara}(para) <- value

\S4method{retcor.profStep}{metaXpara}(para) <- value

\S4method{retcor.plottype}{metaXpara}(para) <- value

\S4method{qcRlscSpan}{metaXpara}(para) <- value

\S4method{xcmsSetObj}{metaXpara}(para) <- value
}
\arguments{
\item{para}{A metaXpara object}

\item{value}{New value}
}
\value{
A object of metaXpara
}
\description{
An S4 class to represent the parameters and data for data processing
}
\section{Methods (by generic)}{
\itemize{
\item \code{dir.case<-}: 

\item \code{dir.ctrl<-}: 

\item \code{sampleListFile<-}: 

\item \code{ratioPairs<-}: 

\item \code{missValueImputeMethod<-}: 

\item \code{outdir<-}: 

\item \code{prefix<-}: 

\item \code{peaksData<-}: 

\item \code{addValueNorm<-}: 

\item \code{rawPeaks<-}: 

\item \code{idres<-}: 

\item \code{xcmsSet.method<-}: 

\item \code{xcmsSet.ppm<-}: 

\item \code{xcmsSet.peakwidth<-}: 

\item \code{xcmsSet.snthresh<-}: 

\item \code{xcmsSet.prefilter<-}: 

\item \code{xcmsSet.mzCenterFun<-}: 

\item \code{xcmsSet.integrate<-}: 

\item \code{xcmsSet.mzdiff<-}: 

\item \code{xcmsSet.noise<-}: 

\item \code{xcmsSet.verbose.columns<-}: 

\item \code{xcmsSet.polarity<-}: 

\item \code{xcmsSet.profparam<-}: 

\item \code{xcmsSet.nSlaves<-}: 

\item \code{xcmsSet.fitgauss<-}: 

\item \code{xcmsSet.sleep<-}: 

\item \code{xcmsSet.fwhm<-}: 

\item \code{xcmsSet.max<-}: 

\item \code{xcmsSet.step<-}: 

\item \code{group.bw0<-}: 

\item \code{group.mzwid0<-}: 

\item \code{group.bw<-}: 

\item \code{group.mzwid<-}: 

\item \code{group.minfrac<-}: 

\item \code{group.minsamp<-}: 

\item \code{group.max<-}: 

\item \code{group.sleep<-}: 

\item \code{retcor.method<-}: 

\item \code{retcor.profStep<-}: 

\item \code{retcor.plottype<-}: 

\item \code{qcRlscSpan<-}: 

\item \code{xcmsSetObj<-}: 
}}
\section{Slots}{

\describe{
\item{\code{dir.case}}{The path names of the NetCDF/mzXML files to read}

\item{\code{dir.ctrl}}{The path names of the NetCDF/mzXML files to read}

\item{\code{sampleListFile}}{The file name of containing the experiment design}

\item{\code{sampleList}}{A data.frame containing the experiment design}

\item{\code{ratioPairs}}{A character containing the ratio pairs, such as "A:B;A:C"}

\item{\code{missValueImputeMethod}}{A character of missing value imputation method}

\item{\code{sampleListHead}}{The name of head of \code{sampleListFile}}

\item{\code{outdir}}{The output directory}

\item{\code{prefix}}{The prefix of output file}

\item{\code{xcmsPeakListFile}}{The file of output from \pkg{XCMS}}

\item{\code{fig}}{A \code{list} of file names of figures}

\item{\code{peaksData}}{A \code{data.frame} containing the peaks data}

\item{\code{VIP}}{A \code{data.frame} containing the VIP}

\item{\code{rawPeaks}}{A \code{data.frame} containg the raw peaks data}

\item{\code{xcmsSetObj}}{An object of \code{\link{xcmsSet}}}

\item{\code{quant}}{A \code{data.frame} containing the quantification result}

\item{\code{idres}}{A \code{data.frame} containing the identification result}

\item{\code{xcmsSet.method}}{Method to use for peak detection. See details
\code{\link{findPeaks}} in package \pkg{XCMS}}

\item{\code{xcmsSet.ppm}}{The maxmial tolerated m/z deviation in consecutive scans,
in ppm (parts per million)}

\item{\code{xcmsSet.peakwidth}}{Chromatographic peak width, given as
range (min,max) in seconds}

\item{\code{xcmsSet.snthresh}}{The signal to noise ratio cutoff, definition see
\code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.prefilter}}{prefilter=c(k,I),
see \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.mzCenterFun}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.integrate}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.mzdiff}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.noise}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.verbose.columns}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.polarity}}{Filter raw data for positive/negative scans.
See \code{\link{xcmsSet}}}

\item{\code{xcmsSet.profparam}}{Parameters to use for profile generation.
See \code{\link{xcmsSet}}}

\item{\code{xcmsSet.nSlaves}}{The number of slaves/cores to be used for parallel
peak detection. See \code{\link{xcmsSet}}}

\item{\code{xcmsSet.fitgauss}}{See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.sleep}}{The number of seconds to pause between plotting
peak finding cycles. See \code{\link{findPeaks.centWave}}}

\item{\code{xcmsSet.fwhm}}{See \code{\link{findPeaks.matchedFilter}}}

\item{\code{xcmsSet.max}}{See \code{\link{findPeaks.matchedFilter}}}

\item{\code{xcmsSet.step}}{See \code{\link{findPeaks.matchedFilter}}}

\item{\code{group.bw0}}{See \code{\link{group.density}}}

\item{\code{group.mzwid0}}{See \code{\link{group.density}}}

\item{\code{group.bw}}{See \code{\link{group.density}}}

\item{\code{group.mzwid}}{See \code{\link{group.density}}}

\item{\code{group.minfrac}}{See \code{\link{group.density}}}

\item{\code{group.minsamp}}{See \code{\link{group.density}}}

\item{\code{group.max}}{See \code{\link{group.density}}}

\item{\code{group.sleep}}{See \code{\link{group.density}}}

\item{\code{retcor.method}}{See \code{\link{retcor}}}

\item{\code{retcor.profStep}}{See \code{\link{retcor.obiwarp}}}

\item{\code{retcor.plottype}}{See \code{\link{retcor.obiwarp}}}

\item{\code{qcRlscSpan}}{The value of span for QC-RLSC}
}}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

