% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{missingValueImpute}
\alias{missingValueImpute}
\alias{missingValueImpute,data.frame-method}
\alias{missingValueImpute,metaXpara-method}
\title{Missing value imputation}
\usage{
missingValueImpute(x, valueID = "value", method = "knn", negValue = TRUE,
  cpu = 1, ...)

\S4method{missingValueImpute}{metaXpara}(x, valueID = "value",
  method = "knn", negValue = TRUE, cpu = 1, ...)

\S4method{missingValueImpute}{data.frame}(x, valueID = "value",
  method = "knn", negValue = TRUE, cpu = 1, ...)
}
\arguments{
\item{x}{The value needed to be imputated}

\item{valueID}{The name of the column which will be used}

\item{method}{Method for imputation: bpca,knn,svdImpute,rf,min}

\item{negValue}{A logical indicates whether convert <=0 value to NA}

\item{cpu}{The number of cpus used}

\item{...}{Additional parameters}
}
\value{
The imputation data
}
\description{
Missing value imputation
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 

\item \code{data.frame}: 
}}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

