% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\docType{methods}
\name{myPLSDA}
\alias{myPLSDA}
\alias{myPLSDA,ANY,ANY,ANY,ANY-method}
\alias{myPLSDA,ANY,ANY,logical,ANY-method}
\alias{myPLSDA,ANY,ANY,missing,logical-method}
\title{Perform PLS-DA analysis}
\usage{
myPLSDA(x, y, save, select, ...)

\S4method{myPLSDA}{ANY,ANY,logical,ANY}(x, y, ncomp = 10, validation = "CV",
  k = 7, method = "oscorespls", save = TRUE)

\S4method{myPLSDA}{ANY,ANY,ANY,ANY}(x, y, ncomp = 10, validation = "CV",
  k = 7, method = "oscorespls")

\S4method{myPLSDA}{ANY,ANY,missing,logical}(x, y, ncomp = 10,
  validation = "CV", k = 7, method = "oscorespls", select = TRUE)
}
\arguments{
\item{x}{A matrix of observations}

\item{y}{a vector or matrix of responses}

\item{save}{A logical indicates whether save the \code{plsr} result}

\item{select}{A logical indicates whether select the best component}

\item{...}{Additional parameters}

\item{ncomp}{The number of component used for PLS-DA}

\item{validation}{See \code{\link{plsr}}}

\item{k}{k-fold}

\item{method}{See \code{\link{plsr}}}
}
\value{
The PLS-DA result
}
\description{
Perform PLS-DA analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{x = ANY,y = ANY,save = logical,select = ANY}: 

\item \code{x = ANY,y = ANY,save = ANY,select = ANY}: 

\item \code{x = ANY,y = ANY,save = missing,select = logical}: 
}}
\examples{
x <- matrix(rnorm(1000),nrow = 10,ncol = 100)
y <- rep(0:1,5)
res <- myPLSDA(x,y,save=TRUE,ncomp=2,validation="CV",k=7,
    method="oscorespls")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

