% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{normalize}
\alias{normalize}
\title{Normalisation of peak intensity}
\usage{
normalize(para, method = "sum", valueID = "value", ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{method}{The normalization method: sum, vsn, quantiles, 
quantiles.robust, sum, pqn. Default is sum.}

\item{valueID}{The name of the column which will be normalized.}

\item{...}{Additional parameter}
}
\value{
A metaXpara object.
}
\description{
The normalize method performs normalisation on peak 
intensities.
}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt", 
    package = "metaX")
para <- reSetPeaksData(para)
para <- metaX::normalize(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

