% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{peakStat}
\alias{peakStat}
\alias{peakStat,metaXpara,plsDAPara-method}
\title{Do the univariate and multivariate statistical analysis}
\usage{
peakStat(para, plsdaPara, doROC = TRUE, saveRds = FALSE, ...)

\S4method{peakStat}{metaXpara,plsDAPara}(para, plsdaPara, doROC = TRUE,
  saveRds = FALSE, ...)
}
\arguments{
\item{para}{A \code{metaXpara} object}

\item{plsdaPara}{A \code{plsDAPara} object}

\item{doROC}{A logical indicates whether to calculate the ROC}

\item{saveRds}{Boolean, setting the argument to TRUE to save some objects to
disk for debug. Only useful for developer. Default is FALSE.}

\item{...}{Additional parameter}
}
\value{
none

An object of \code{metaXpara}
}
\description{
Do the univariate and multivariate statistical analysis
}
\section{Methods (by class)}{
\itemize{
\item \code{para = metaXpara,plsdaPara = plsDAPara}: 
}}
\examples{
\dontrun{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
ratioPairs(para) <- "S:C"
addValueNorm(para) <- para
plsdaPara <- new("plsDAPara")
res <- peakStat(para,plsdaPara)
}
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

