% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotCV}
\alias{plotCV}
\alias{plotCV,metaXpara-method}
\title{Plot the CV distribution of peaks in each group}
\usage{
plotCV(x, ...)

\S4method{plotCV}{metaXpara}(x, ...)
}
\arguments{
\item{x}{A metaXpara object}

\item{...}{Additional parameter}
}
\value{
none
}
\description{
Plot the CV distribution of peaks in each group.
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 
}}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt",
    package = "metaX")
para <- reSetPeaksData(para)
plotCV(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

