% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotMissValue}
\alias{plotMissValue}
\alias{plotMissValue,metaXpara-method}
\title{Plot missing value distribution}
\usage{
plotMissValue(para, width = 8, height = 5, ...)

\S4method{plotMissValue}{metaXpara}(para, width = 8, height = 5, ...)
}
\arguments{
\item{para}{A metaXpara object}

\item{width}{The width of the graphics region in inches.
The default values are 8.}

\item{height}{The height of the graphics region in inches.
The default values are 5.}

\item{...}{Additional parameter}
}
\value{
The figure name
}
\description{
Plot missing value distribution.
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 
}}
\examples{
library(faahKO)
xset <- group(faahko)
xset <- retcor(xset)
xset <- group(xset)
xset <- fillPeaks(xset)
peaksData <- as.data.frame(groupval(xset,"medret",value="into"))
peaksData$name <- row.names(peaksData)
para <- new("metaXpara")
rawPeaks(para) <- peaksData
sampleListFile(para) <- system.file("extdata/faahKO_sampleList.txt",
    package = "metaX")
para <- reSetPeaksData(para)
plotMissValue(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

