% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tool.R
\docType{methods}
\name{plotPeakSumDist}
\alias{plotPeakSumDist}
\alias{plotPeakSumDist,metaXpara-method}
\title{Plot the total peak intensity distribution}
\usage{
plotPeakSumDist(para, valueID = "value", width = 6, height = 4, ...)

\S4method{plotPeakSumDist}{metaXpara}(para, valueID = "value", width = 6,
  height = 4, ...)
}
\arguments{
\item{para}{A metaXpara object.}

\item{valueID}{The name of the column used}

\item{width}{Width of the figure}

\item{height}{Height of the figure}

\item{...}{Other argument}
}
\value{
The output figure name.
}
\description{
Plot the total peak intensity distribution
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 
}}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
plotPeakSumDist(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

