% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{plotQCRLSC}
\alias{plotQCRLSC}
\alias{plotQCRLSC,metaXpara-method}
\title{Plot figures for QC-RLSC}
\usage{
plotQCRLSC(para)

\S4method{plotQCRLSC}{metaXpara}(para)
}
\arguments{
\item{para}{A \code{metaXpara} object}
}
\value{
none
}
\description{
Plot figures for QC-RLSC
}
\section{Methods (by class)}{
\itemize{
\item \code{metaXpara}: 
}}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)[1:20,]
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
res <- doQCRLSC(para,cpu=1)
plotQCRLSC(res$metaXpara)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
\seealso{
\code{\link{doQCRLSC}}
}

