% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metaX.R
\docType{class}
\name{plsDAPara-class}
\alias{center<-,plsDAPara-method}
\alias{kfold<-,plsDAPara-method}
\alias{method<-,plsDAPara-method}
\alias{ncomp<-,plsDAPara-method}
\alias{nperm<-,plsDAPara-method}
\alias{plsDAPara-class}
\alias{scale<-,plsDAPara-method}
\alias{t<-,plsDAPara-method}
\alias{validation<-,plsDAPara-method}
\title{An S4 class to represent the parameters for PLS-DA analysis}
\usage{
\S4method{scale}{plsDAPara}(para) <- value

\S4method{center}{plsDAPara}(para) <- value

\S4method{t}{plsDAPara}(para) <- value

\S4method{validation}{plsDAPara}(para) <- value

\S4method{ncomp}{plsDAPara}(para) <- value

\S4method{nperm}{plsDAPara}(para) <- value

\S4method{kfold}{plsDAPara}(para) <- value

\S4method{method}{plsDAPara}(para) <- value
}
\arguments{
\item{para}{A metaXpara object}

\item{value}{New value}
}
\value{
A object of plsDAPara
}
\description{
An S4 class to represent the parameters for PLS-DA analysis
}
\section{Methods (by generic)}{
\itemize{
\item \code{scale<-}: 

\item \code{center<-}: 

\item \code{t<-}: 

\item \code{validation<-}: 

\item \code{ncomp<-}: 

\item \code{nperm<-}: 

\item \code{kfold<-}: 

\item \code{method<-}: 
}}
\section{Slots}{

\describe{
\item{\code{scale}}{The method used to scale the data,
see \code{\link{preProcess}} in \pkg{metaX}}

\item{\code{center}}{A logical which indicates if the matrix should be mean
centred or not}

\item{\code{t}}{The method used to transform the data,
see \code{\link{transformation}} in \pkg{metaX}}

\item{\code{validation}}{The method for validation, default is "CV"}

\item{\code{ncomp}}{The number of components used for PLS-DA, default is 2}

\item{\code{nperm}}{The number of permutations, default is 200}

\item{\code{kfold}}{The number of folds for cross-validation, default is 7}

\item{\code{do}}{A logical which indicates whether to do the plsDA analysis,
default is TRUE}

\item{\code{method}}{The method used in PLS-DA. See \code{\link{plsr}}
in \pkg{pls}}
}}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

