% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prep.R
\name{preProcess}
\alias{preProcess}
\title{Pre-Processing}
\usage{
preProcess(para, log = FALSE, scale = c("none", "pareto", "vector", "uv"),
  center = TRUE, valueID = "valueNorm", ...)
}
\arguments{
\item{para}{An metaX object}

\item{log}{A logical indicates whether do the log transformation}

\item{scale}{The method of scaling}

\item{center}{Centering}

\item{valueID}{The name of column used for transformation}

\item{...}{Additional parameter}
}
\value{
An new metaX object
}
\description{
Pre-Processing
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
para <- missingValueImpute(para)
para <- preProcess(para,valueID = "value",scale="uv")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

