% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaX.R
\docType{methods}
\name{reSetPeaksData}
\alias{reSetPeaksData}
\title{reSetPeaksData}
\usage{
reSetPeaksData(para)
}
\arguments{
\item{para}{An object of \code{metaXpara}}
}
\value{
none
}
\description{
reSetPeaksData
}
\examples{
para <- new("metaXpara")
pfile <- system.file("extdata/MTBLS79.txt",package = "metaX")
sfile <- system.file("extdata/MTBLS79_sampleList.txt",package = "metaX")
rawPeaks(para) <- read.delim(pfile,check.names = FALSE)
sampleListFile(para) <- sfile
para <- reSetPeaksData(para)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

