% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{runOPLSDA}
\alias{runOPLSDA}
\title{run OPLS-DA}
\usage{
runOPLSDA(para, oplsdaPara, auc = TRUE, sample = NULL,
  valueID = "valueNorm", label = "order", ...)
}
\arguments{
\item{para}{An object of metaXpara}

\item{oplsdaPara}{An object of oplsDAPara}

\item{sample}{Sample class}

\item{valueID}{The name of column used}

\item{label}{The label used for plot}

\item{...}{additional arguments}
}
\value{
A list object
}
\description{
Perform the OPLS-DA analysis
}

